general restructure
