+++ crates/cli/commands/Cargo.toml
@@ -64,9 +64,6 @@ ratatui = { version = "0.27", default-features = false, features = [
     "crossterm",
 ] }
 
-# metrics
-metrics-process.workspace = true
-
 # reth test-vectors
 proptest = { workspace = true, optional = true }
 arbitrary = { workspace = true, optional = true }
