+++ crates/node/metrics/src/lib.rs
@@ -7,6 +7,8 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+/// The metrics hooks for prometheus.
+pub mod hooks;
 pub mod recorder;
 /// The metric server serving the metrics.
 pub mod server;
